overlays: Adjust for the new snd_bcm2835 handling
The snd_bcm2835 driver has lost its instantiation by Device Tree,
requiring a few adjustments to overlays, in particular getting
them to append module parameters to the kernel command line.

N.B. A consequence of this change is that runtime application of
overlays that change the onboard audio handling is unlikely to work
as expected.

Signed-off-by: Phil Elwell <phil@raspberrypi.com>